Rollback centering spatial search widget on Meerbusch
There's a better approach to centering the widget: altering the
same template in the viderum spatial extension to accept a default
extent and then grabbing that variable from the portal config.
That way, the map for any site using the widget can be centered
just by updating the config rather than adding a new template.